refactor: Create Attachment component
기존 컨테이너에 첨부파일 GUI를 포함하는 형태에서 컴포넌트를 분리하였습니다.
